+++ src/sink.rs
@@ -119,13 +119,13 @@ impl Sink {
 
         let source = source
             .speed(1.0)
-            // must be placed before pausable but after speed & delay
+            // Must be placed before pausable but after speed & delay
             .track_position()
             .pausable(false)
             .amplify(1.0)
             .skippable()
             .stoppable()
-            // if you change the duration update the docs for try_seek!
+            // If you change the duration update the docs for try_seek!
             .periodic_access(Duration::from_millis(5), move |src| {
                 if controls.stopped.load(Ordering::SeqCst) {
                     src.stop();
@@ -177,6 +177,14 @@ impl Sink {
         *self.controls.volume.lock().unwrap() = value;
     }
 
+    /// Gets the speed of the sound.
+    ///
+    /// See [`Sink::set_speed`] for details on what *speed* means.
+    #[inline]
+    pub fn speed(&self) -> f32 {
+        *self.controls.speed.lock().unwrap()
+    }
+
     /// Changes the play speed of the sound. Does not adjust the samples, only the playback speed.
     ///
     /// # Note:
@@ -190,24 +198,6 @@ impl Sink {
     /// - If you set the speed to 2 the total duration will be halve of what it
     ///   was.
     ///
-    /// See [`Speed`](crate::source::Speed) for details
-    #[inline]
-    pub fn speed(&self) -> f32 {
-        *self.controls.speed.lock().unwrap()
-    }
-
-    /// Changes the speed of the sound.
-    ///
-    /// The value `1.0` is the "normal" speed (unfiltered input). Any value other than `1.0` will
-    /// change the play speed of the sound.
-    ///
-    /// #### Note:
-    /// 1. **Increasing the speed would also increase the pitch by the same factor**
-    /// - If you increased set the speed to 0.5, the frequency would be slower (0.5x the original frequency) .
-    /// - Also if you set the speed to 1.5 the frequency would be faster ( 1.5x the original frequency).
-    /// 2. **Change in the speed would affect your total duration inversely**
-    /// - if you set the speed by 0.5, your total duration would be (2x the original total duration) longer.
-    /// - Also if you set the speed to 2 the total duration would be (0.5 the original total_duration) shorter
     #[inline]
     pub fn set_speed(&self, value: f32) {
         *self.controls.speed.lock().unwrap() = value;