+++ src/stream.rs
@@ -1,3 +1,10 @@
+//! Output audio via the OS via mixers or play directly
+//!
+//! This module provides a builder that's used to configure and open audio output. Once
+//! opened sources can be mixed into the output via `OutputStream::mixer`.
+//!
+//! There is also a convenience function `play` for using that output mixer to
+//! play a single sound.
 use crate::common::{ChannelCount, SampleRate};
 use crate::decoder;
 use crate::mixer::{mixer, Mixer, MixerSource};
@@ -14,6 +21,7 @@ const HZ_44100: SampleRate = 44_100;
 /// Use `mixer()` method to control output.
 /// If this is dropped, playback will end, and the associated output stream will be disposed.
 pub struct OutputStream {
+    config: OutputStreamConfig,
     mixer: Mixer,
     _stream: cpal::Stream,
 }
@@ -23,10 +31,14 @@ impl OutputStream {
     pub fn mixer(&self) -> &Mixer {
         &self.mixer
     }
+
+    pub fn config(&self) -> &OutputStreamConfig {
+        &self.config
+    }
 }
 
 #[derive(Copy, Clone, Debug)]
-struct OutputStreamConfig {
+pub struct OutputStreamConfig {
     channel_count: ChannelCount,
     sample_rate: SampleRate,
     buffer_size: BufferSize,
@@ -41,6 +53,24 @@ impl Default for OutputStreamConfig {
             buffer_size: BufferSize::Default,
             sample_format: SampleFormat::F32,
         }
+    }
+}
+
+impl OutputStreamConfig {
+    pub fn channel_count(&self) -> ChannelCount {
+        self.channel_count
+    }
+
+    pub fn sample_rate(&self) -> SampleRate {
+        self.sample_rate
+    }
+
+    pub fn buffer_size(&self) -> &BufferSize {
+        &self.buffer_size
+    }
+
+    pub fn sample_format(&self) -> SampleFormat {
+        self.sample_format
     }
 }
 
@@ -412,6 +442,7 @@ impl OutputStream {
             Ok(Self {
                 _stream: stream,
                 mixer: controller,
+                config: config.clone(),
             })
         })
     }
@@ -542,7 +573,7 @@ impl OutputStream {
 }
 
 /// Return all formats supported by the device.
-fn supported_output_configs(
+pub fn supported_output_configs(
     device: &cpal::Device,
 ) -> Result<impl Iterator<Item = cpal::SupportedStreamConfig>, StreamError> {
     let mut supported: Vec<_> = device